backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djoser==2.2.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.5.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
django-colorfield==0.9.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
python-slugify==8.0.1
pdfrw==0.4
pandas==2.0.2